requirements.txt
supabase==2.5.1
streamlit==1.37.1
python-dotenv==1.0.1
# Streamlit for UI
streamlit==1.37.1

# Supabase client and dependencies
supabase==2.5.1
supabase-auth==2.5.1
supabase-functions==2.5.1
storage3==2.21.1
postgrest==2.21.1
httpx==0.27.0
cryptography==46.0.2
cffi==2.0.0
pyjwt==2.10.1
pycparser==2.23

# Realtime client (compatible with supabase 2.5.1)
realtime==2.21.0

# Environment variables
python-dotenv==1.0.1

# Optional: Type hints and utilities
typing-extensions==4.15.0
anyio==4.11.0
